# swagger.yml
openapi: 3.0.2
info:
  title: Medical Database
  version: 1.0.0
  description: health_one_medical_database
  termsOfService: 'http://swagger.io/terms/'
paths:
  /doctors:
    get:
      tags:
        - /doctors
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/doctor'
              examples:
                doctor:
                  value:
                    Id: 1
                    specialty: Hambaarst
                    hospitals: 2
                    phoneNumber: 57743229
                    address: Pärna 8
          description: OK
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/doctor'
              examples:
                doctor:
                  value: 'error: "No doctors currently exists!"'
          description: Not Found
        '500':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/doctor'
              examples:
                doctor:
                  value: 'message: "Could not fetch doctors!"'
          description: Internal Server Error
      summary: Get all doctors
      description: Gets all the doctors from the database.
    post:
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/doctor'
            examples:
              doctor:
                value:
                  specialty: hambaarst
                  hospitals: 1
                  phoneNumber: 56654334
                  address: Pärna 8
        required: true
      tags:
        - /doctors
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/doctor'
              examples:
                doctor:
                  value:
                    specialty: kirurg
                    hospitals: 1
                    phoneNumber: 56654334
                    address: Pärna 8
                    Id: 3
          description: OK
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/doctor'
              examples:
                doctor:
                  value: >-
                    error: "Doctor has to have a valid specialty, hospitals, phone number
                    and address!"
          description: Bad Request
        '500':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/doctor'
              examples:
                doctor:
                  value: 'message: "Could not fetch doctors!"'
          description: Internal Server Error
      summary: Add a doctor
...
components:
  schemas:
    doctor:
      title: Root Type for doctor
      description: ''
      required:
        - address
        - specialty
        - hospitals
        - Id
        - phoneNumber
      type: object
      properties:
        Id:
          maximum: 10
          type: integer
          example: |-
            {
                "Id": 1
            }
        specialty:
          maxLength: 50
          type: string
          example: |-
            {
                "specialty": "hambaarst"
            }
        hospitals:
          maximum: 10
          type: integer
          example: |-
            {
                "hospitals": 2
            }
        phoneNumber:
          maximum: 10
          type: integer
          example: |-
            {
                "phoneNumber": 57791345
            }
        address:
          maxLength: 255
          type: string
          example: |-
            {
                "address": "Pärna 8"
            }
      example:
        Id: 1
        specialty: Hambaarst
        hospitals: 2
        phoneNumber: 54482009
        address: Pärna 5
...